OpenSearch refactor.
