+++ .env_file.example
@@ -1,5 +1,5 @@
 IMAGE_MODEL=phash
-ELASTICSEARCH_URL=http://elasticsearch:9200
+OPENSEARCH_URL=http://elasticsearch:9200
 REDIS_HOST=redis
 REDIS_PORT=6379
 REDIS_DATABASE=0
