+++ .travis.yml
@@ -12,7 +12,7 @@ before_script:
 - docker-compose build --pull
 - docker-compose -f docker-compose.yml -f docker-test.yml up -d
 - docker-compose logs -t -f &
-- echo "Waiting for Elasticsearch indexes..." && until curl --silent --fail -I "http://localhost:9200/alegre_similarity_test"; do sleep 1; done
+- echo "Waiting for OpenSearch indexes..." && until curl --silent --fail -I "http://localhost:9200/alegre_similarity_test"; do sleep 1; done
 - until curl --silent --fail -I "http://localhost:3100"; do sleep 1; done
 - echo "Waiting for model servers..." && while [[ ! '2' =~ $(redis-cli -n 1 SCARD 'SharedModel') ]]; do sleep 1; done
 #comment until fix timeout curl: (28) Operation timed out
