+++ app/main/controller/similarity_controller.py
@@ -9,12 +9,12 @@ api = Namespace('similarity', description='text similarity operations')
 similarity_request = api.model('similarity_request', {
     'text': fields.String(required=False, description='text to be stored or queried for similarity'),
     'doc_id': fields.String(required=False, description='text ID to constrain uniqueness'),
-    'model': fields.String(required=False, description='similarity model to use: "elasticsearch" (pure Elasticsearch, default) or the key name of an active model'),
-    'models': fields.List(required=False, description='similarity models to use: ["elasticsearch"] (pure Elasticsearch, default) or the key name of an active model', cls_or_instance=fields.String),
+    'model': fields.String(required=False, description='similarity model to use: "opensearch" (pure Elasticsearch, default) or the key name of an active model'),
+    'models': fields.List(required=False, description='similarity models to use: ["opensearch"] (pure Elasticsearch, default) or the key name of an active model', cls_or_instance=fields.String),
     'language': fields.String(required=False, description='language code for the analyzer to use during the similarity query (defaults to standard analyzer)'),
     'threshold': fields.Float(required=False, description='minimum score to consider, between 0.0 and 1.0 (defaults to 0.9)'),
     'context': JsonObject(required=False, description='context'),
-    'fuzzy': fields.Boolean(required=False, description='whether or not to use fuzzy search on GET queries (only used when model is set to \'elasticsearch\')'),
+    'fuzzy': fields.Boolean(required=False, description='whether or not to use fuzzy search on GET queries (only used when model is set to \'opensearch\')'),
 })
 @api.route('/')
 class SimilarityResource(Resource):
@@ -39,7 +39,7 @@ class SimilarityResource(Resource):
         return similarity.add_item(item, "text")
 
     @api.response(200, 'text similarity successfully queried.')
-    @api.doc('Make a text similarity query. Note that we currently require GET requests with a JSON body rather than embedded params in the URL. You can achieve this via curl -X GET -H "Content-type: application/json" -H "Accept: application/json" -d \'{"text":"Some Text", "threshold": 0.5, "model": "elasticsearch"}\' "http://[ALEGRE_HOST]/text/similarity"')
-    @api.doc(params={'text': 'text to be stored or queried for similarity', 'threshold': 'minimum score to consider, between 0.0 and 1.0 (defaults to 0.9)', 'model': 'similarity model to use: "elasticsearch" (pure Elasticsearch, default) or the key name of an active model'})
+    @api.doc('Make a text similarity query. Note that we currently require GET requests with a JSON body rather than embedded params in the URL. You can achieve this via curl -X GET -H "Content-type: application/json" -H "Accept: application/json" -d \'{"text":"Some Text", "threshold": 0.5, "model": "opensearch"}\' "http://[ALEGRE_HOST]/text/similarity"')
+    @api.doc(params={'text': 'text to be stored or queried for similarity', 'threshold': 'minimum score to consider, between 0.0 and 1.0 (defaults to 0.9)', 'model': 'similarity model to use: "opensearch" (pure Elasticsearch, default) or the key name of an active model'})
     def get(self):
       return similarity.get_similar_items(similarity.get_body_for_text_document(request.args or request.json, mode='query'), "text")
