--- elasticsearch/Dockerfile
@@ -1,3 +1,0 @@
-FROM docker.elastic.co/elasticsearch/elasticsearch:7.9.2
-RUN echo y | bin/elasticsearch-plugin install analysis-icu
-RUN echo y | bin/elasticsearch-plugin install repository-s3
