--- elasticsearch/alegre_similarity_base.json
@@ -1,10 +1,0 @@
-{
-  "properties": {
-    "content": {
-      "type": "text"
-    },
-    "context": {
-      "type": "nested"
-    }
-  }
-}
